requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
requests